Renamed sky_remote folder
intent is for functionality beyond just a remote, also media player
